+++ resolwe/storage/management/__init__.py
@@ -1,0 +1,10 @@
+""".. Ignore pydocstyle D400.
+
+==================
+Storage Management
+==================
+
+.. automodule:: resolwe.storage.management.commands.run_storage_manager
+    :members:
+
+"""