+++ docs/CHANGELOG.rst
@@ -21,6 +21,7 @@ Added
   service and local filesystem.
 - Add migrations to move from old storage model to the new one
 - Add storage manager
+- Add management command to start storage manager
 
 ===================
 21.1.0 - 2020-04-14